added jcr services
