+++ services/pom.xml
@@ -61,8 +61,6 @@
       <version>1.7.28</version>
     </dependency>
 
-
-
   </dependencies>
   <name>Training Exercise - Services</name>
   <build>
@@ -71,7 +69,7 @@
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-surefire-plugin</artifactId>
-        <version>2.22.1</version>
+
       </plugin>
     </plugins>
   </build>
