+++ services/src/main/java/org/exoplatform/addons/services/Hanen_StartableService.java
@@ -9,7 +9,8 @@ public class Hanen_StartableService implements Startable{
 	 private static final Log LOG = ExoLogger.getExoLogger(Hanen_StartableService.class);
 
 	@Override
-	public  void start() {		 LOG.info("loging in from Hanen_Startable Service !");
+	public  void start() {
+		 LOG.info("loging in from Hanen_Startable Service !");
 		 PortalContainer portalContainer = (PortalContainer) ExoContainerContext.getCurrentContainer();
 		 Hanen_Service myNewService = (Hanen_Service) portalContainer.getComponentInstanceOfType(Hanen_Service.class);
 		 LOG.info(myNewService.call(" Calling function call of Hanen_Service !"));
@@ -17,6 +18,7 @@ public class Hanen_StartableService implements Startable{
 
 	@Override
 	public void stop() {
+
 		 LOG.info("loging out from Hanen_Startable Service !");
 	}
 }
